Add files via upload
Based on the outputs of the different ML methods, it can be observed that the Decision Tree method has the highest accuracy, precision, and F1 score. The K-Nearest Neighbor method has a lower accuracy, precision, and F1 score compared to the Multinomial Logistic Regression and HPCI. In contrast, the Random Forest predictor used by Shreyas Kharasambale has a precision, recall, and F1 score higher than the Multinomial Logistic Regression, but lower than the Decision Tree.
Overall, it is important to note that the choice of ML method is dependent on various factors such as the size and complexity of the dataset, the specific problem being solved, and the resources available. It is recommended to compare multiple methods and evaluate their performance to determine the best fit for a given problem.